feat(ai-ready): G11 evaluation and certification loop, acceptance 5/6
- rag evaluation live: BM25 five metrics reproducible remotely
  (Recall@5 1.0 / MRR 0.95 / Citation 0.9 / Faithfulness 0.3, 10-question
  set anchored to real chunk fingerprints)
- certification workflow: V11 request table (DOUBLE PRECISION - plain
  DOUBLE breaks real PostgreSQL), submit guarded (ASSESSED-only, fixed a
  real defect where CERTIFIED products could re-submit), decision
  restricted to admin roles, direct CERTIFIED transition rejected 409;
  remote run proved PENDING -> APPROVED -> CERTIFIED
- evaluate endpoint merges metrics into version readiness_json
- om-sync script: glossary + AIReadiness classification/tags delivered;
  term write blocked by an instance-level glossary-reference defect
  (documented, backlog P3 evidence upgraded, script degrades gracefully)
- portal: evaluation panel + certification actions/history

+++ services/control-plane/src/main/resources/db/migration/V11__ai_certification_request.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS data_os.ai_certification_request (
     id                VARCHAR(36)  PRIMARY KEY,
     product_id        VARCHAR(36)  NOT NULL,
     version_sn        VARCHAR(32)  NOT NULL,
-    readiness_overall DOUBLE       NOT NULL,
+    readiness_overall DOUBLE PRECISION NOT NULL,
     certification     VARCHAR(24)  NOT NULL,
     decision          VARCHAR(16)  NOT NULL DEFAULT 'PENDING',
     decision_note     VARCHAR(500),
